+++ src/main/java/com/raoulvdberge/refinedstorage/tile/grid/TileGrid.java
@@ -24,19 +24,19 @@ public class TileGrid extends TileNode<NetworkNodeGrid> {
             t.getNode().setViewType(v);
             t.getNode().markDirty();
         }
-    }, p -> GuiGrid.markForSorting());
+    }, p -> GuiGrid.scheduleSort());
     public static final TileDataParameter<Integer, TileGrid> SORTING_DIRECTION = new TileDataParameter<>(DataSerializers.VARINT, 0, t -> t.getNode().getSortingDirection(), (t, v) -> {
         if (IGrid.isValidSortingDirection(v)) {
             t.getNode().setSortingDirection(v);
             t.getNode().markDirty();
         }
-    }, p -> GuiGrid.markForSorting());
+    }, p -> GuiGrid.scheduleSort());
     public static final TileDataParameter<Integer, TileGrid> SORTING_TYPE = new TileDataParameter<>(DataSerializers.VARINT, 0, t -> t.getNode().getSortingType(), (t, v) -> {
         if (IGrid.isValidSortingType(v)) {
             t.getNode().setSortingType(v);
             t.getNode().markDirty();
         }
-    }, p -> GuiGrid.markForSorting());
+    }, p -> GuiGrid.scheduleSort());
     public static final TileDataParameter<Integer, TileGrid> SEARCH_BOX_MODE = new TileDataParameter<>(DataSerializers.VARINT, 0, t -> t.getNode().getSearchBoxMode(), (t, v) -> {
         if (IGrid.isValidSearchBoxMode(v)) {
             t.getNode().setSearchBoxMode(v);
@@ -62,7 +62,7 @@ public class TileGrid extends TileNode<NetworkNodeGrid> {
         t.getNode().markDirty();
     }, p -> {
         if (Minecraft.getMinecraft().currentScreen instanceof GuiGrid) {
-            GuiGrid.markForSorting();
+            GuiGrid.scheduleSort();
         }
     });
     public static final TileDataParameter<Integer, TileGrid> TAB_PAGE = new TileDataParameter<>(DataSerializers.VARINT, 0, t -> t.getNode().getTabPage(), (t, v) -> {